Add - remove refactor
