+++ index.html
@@ -43,6 +43,6 @@
     </ul>
   </section>
   
-  <script type="module" src="index.js"></script>
+  <script type="module" src="./modules/index.js"></script>
 </body>
 </html>
