+++ modules/index.js
@@ -1,21 +1,9 @@
-// variables
-const addButton = document.getElementById('addButton');
-const title = document.getElementById('title');
-const author = document.getElementById('author');
-
-const list = document.getElementById('list'); // link LIST
-const addBook = document.getElementById('addBook');
-
-const addNew = document.getElementById('addNew'); // link ADDNEW
-const newBook = document.getElementById('newBook');
-
-const contact = document.getElementById('contact'); // link CONTACT
-const contactBook = document.getElementById('contactBook');
-
-// imports
-import { addbook } from './add_book.js';
-import {removebook} from './remove-book.js';
+import {drko} from './add_book.js';
+import  {
+  addButton, title, author, list, addBook, addNew, newBook, contact, contactBook
+} from './variable.js';
 import * as stylings from './stylings.js';
+drko();
 
 /* addBook - hold books after every refresh page */
 if (localStorage.getItem('books') !== null) {
@@ -31,24 +19,3 @@ if (localStorage.getItem('books') !== null) {
   });
 }
 /* ***** */
-
-class Methods {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-}
-
-addButton.addEventListener('click', (e) => {
-  e.preventDefault();
-
-  /* add book with method class */
-  const ui = new Methods(title.value, author.value);
-  ui.addbook();
-
-  /* remove book with method class */
-  ui.removebook();
-
-  title.value = '';
-  author.value = '';
-});
