+++ modules/variable.js
@@ -10,3 +10,7 @@ const newBook = document.getElementById('newBook');
 
 const contact = document.getElementById('contact'); // link CONTACT
 const contactBook = document.getElementById('contactBook');
+
+export  {
+    addButton, title, author, list, addBook, addNew, newBook, contact, contactBook
+  };
